maj "coding style" n°5

+++ tools/del_map.py
@@ -1,8 +1,8 @@
 import os
-from glob import glob
 
 path = os.path.join("..", "saves", "map.umd")
 
-os.remove(path)
+if os.path.exists(path):
+    os.remove(path)
 
 print("Map deleted")
